Refactor Context Api (#239)
